+++ tests/integration/openebs/static/static.go
